RBAC: split dashboardBuilder so client-bundle imports resolve
The dev build was crashing with 'dashboardLoader is not a function'
on first navigation to any /admin route, then the browser would
hard-reload back to the previous page. Symptom: clicking 'Admin
dashboard' (or anywhere /@ → /admin chain) flashed admin then bounced
back, with no obvious cause server-side (every loader returned 200).

Root cause: routes export their loader at module top-level via the
wrapper:
    export const loader = dashboardLoader(...);
The factory call evaluates at module load. dashboardBuilder lived in
a .server.ts file, which Remix strips from the client bundle. In the
prod build the loader export + its RHS are both tree-shaken, so the
import is unreferenced and removed — fine. In the dev build the call
is preserved (HMR/source-map friendliness) and resolves
dashboardLoader to undefined on the client, throwing on module load.
Remix's recovery is to reload the page, which lands on the previous
URL because that's the last known-good navigation entry.

Fix: split the wrapper so the import target exists on both server
and client.
- dashboardBuilder.ts (no .server) — exports types + dashboardLoader /
  dashboardAction wrappers. Wrappers return closures whose bodies
  dynamic-import the server impl. The closure body never runs on the
  client, so the dynamic import only resolves at server runtime.
  Client just sees a function that returns another function — the
  top-level call now works there.
- dashboardBuilder.server.ts — slimmed down to authenticateAndAuthorize
  + the redirect/authorization helpers. Imported via dynamic import
  from the wrapper. Stays out of the client bundle.

Routes drop the .server suffix on the import path. No change to the
route's loader/action surface. Verified end-to-end via Chrome
DevTools: /@ → /admin chain renders the admin dashboard cleanly,
no console errors, no extra document fetch back to the org URL.
